requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]>=4.6
scikit-learn

# -e .